refactor(storage): isolate durable engine opening
